LPD-17027 Moving to jaxrs.serializer
